# Add a generic chunk pool for batch.NewChunkMergeIterator