@codesome 2023-05-31
I converted this PR to a draft since from the above comment thread there seem to be multiple different ways this can be done. Need to determine which is better here. This PR is not my top priority at the moment, so I will come back to this sometime later (but if anyone wants to pursue this further, feel free!)